+++ Assets/Game/Scenes/GameScene/NavMesh-Cube 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 251891eed7344ff4e82c6af558652560
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
