+++ Assets/Game/Scripts/Enemies/SeekerPoints.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a51ccb6c3da77ee4484dffd048d88a21
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
