--- Assets/Photon/FusionAddons/Photon-Fusion-Physics-Addon-2.0.0-RC3-807.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8b9ed61988db5114eb0062cbe1959387
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
