--- Assets/Photon/FusionAddons/Photon-Fusion-TestSuit-Addon-2.0.0-RC3-807.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e8e485ea5fdb85e4a80d554bae0ad087
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
