--- Assets/Photon/FusionDemos/Photon-Fusion-Intro-Sample-2.0.0-RC3-807.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5a77421e106bc7a4d8a5d79f2e4c1b7e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
